Hoisted update-array
